eval: jobget()
Deprecate "pty" and "rpc" job options in favor a new "type" option whose value
may be "pty", "rpc", "term", or "".

"term" job-type will be used later to start a pty job with in the terminal
emulator. This needs a new "out_buf" job option which allows a job to control
a buffer.