screen.c: column width returned by function
This commit does not change the behavior but change in column size centralized so it opens up possibilities (like adjusting signcolumn width depending on the maximum number of signs on a line).

The returned value is limited by the size of the `extra` tab in win_line
(currently allows for 18 ANSII characters).