ex_docmd.c: Remove Unix check for multi file args for :edit
Remove the Unix check for more than one file for :edit. "E172: Only one file
name allowed" was thrown when a non-escaped space was detected in the line.
This unifies the default behaviour of :edit to only accept one file, and try
to treat the rest of the line as a file name when encountering a space.

https://github.com/neovim/neovim/issues/6010