terminal: fixes for 'scrollback' option
In new terminals the global value should be used as the new local value
for the option.

In normal buffers, the value is always kept at '-1'. Setting the
global value is allowed using either :set or :setglobal in any type of
buffer. This allows setting the option in an init.vim. :setlocal in
non-terminal buffers raises an error.

Ref #6337.