Windows: shellescape() jobstart test commands
With the change in process argument quoting for system('...') in
Windows, calling powershell via system() with whitespaces inside the
command produces unexpected results.

The jobstart tests use powershell in Windows and need to be changed
accordingly, otherwise the argument string is treated as multiple
arguments and the output is different from expected.

fixup