Use 'exiting' variable to mark while() loop extent
This is used instead of having a condition on the 'c' variable.
We do this so it's clear at what point the while () loop can exit, e.g.
you can search for 'exiting = ' to find all places where it can happen.

Moreover, in the future we may be able to separate concerns between 'c'
as a variable to determine exiting the loop, as a temporary variable,
and as the value to return.

Remove vgetorpeek() unnecessary setting of c = -1

Before it was used simply to signify continuing on in the loop, it's not
needed any more.

'c' is always reset on each iteration around the loop, any value
leftover from the previous loop will be unused.