vgetorpeek swap `break` for `continue`
The difference between using `break` and `continue` is that `break`
puts you into the outer while () loop that checks the typeahead_char
and readbuffers before going back into the main for(;;) loop handling
the typeahead buffer.

This change is valid if there is no benefit to checking the stuffbuffers
and typeahead_char in the places where `break` used to be.
I feel comfortable asserting this is the case, as the methods that the
stuffbuffers can be added to are known and very limited.
The only way the stuffbuffer or typeahead_char can be modified are in
user code evaluating an <expr> mapping (there is no other place in the
vgetorpeek() for(;;) loop that could affect them).
The only way that these hypothetical characters in the stuffbuffer would
be used before characters in the typeahead before this change is if
there were an error looking for characters in the typeahead (either an
integer overflow in the typeahead buffer), an error in the evaluation of
an <expr> mapping, or a mapping that is nested too deeply.

During this evaluation both textlock and ex_normal_lock are both set,
which limits the available code paths.
There are three possible places taht

There are three possible places that can change either the
typeahead_char or the stuffbufers while textlock and ex_normal_lock are
both set.

1) do_more_prompt()
    This sets the typeahead_char when the key ':' is pressed at a
    -- More -- prompt.
    When this is pressed, got_int is also set, which means that this is
    the most likely difference to be observed.
    In order to see this, start vim in a short terminal (i.e. not many
    lines), source the vimrc
        function TestMapping()
          let oldkeymap = &keymap
          set keymap=dvorak
          lmap
          return 'j'
          let &keymap = oldkeymap
        endfunction
        nmap <expr> ab TestMapping()
    press `ab`, see the -- More -- prompt, and press ':'.
    Before this commit you'd get the 'Type  :quit<Enter>  to exit Nvim'
    message, and the map isn't used.
    After this commit the map still isn't used, but you get to the ':'
    prompt instead of getting the 'Type  :quit<Enter>  to exit Nvim'
    message.

2) get_number() and get_user_input()
    These put text into the stuffbuffer.
    In order for the difference between using `continue` and `break` to be
    observed, eval0() must return FAIL.
    This requires some error to have happened, as an error must have happened,
    emsg() is called (either in something lower than eval0(), or eval0()
    itself). emsg() calls flush_buffers() (either indirectly or directly),
    unless emsg_off is set.
    When flush_buffers() is set, both the stuffbuffers and typebuf are cleared,
    so this change doesn't affect anything.