Store 'langmap' adjusted keys in macros
We put LANGMAP_ADJUST() in vgetc() instead of vgetorpeekc(), because
vgetc() is the first place that vim switches from handling bytes to
characters.
The only place vgetorpeek() is called outside of vgetc() is vpeekc()
Each of the calling places of this is written below, along with the
reason that not applying LANGMAP in that case is fine.

insertchar() (in INSERT mode)
command_line_execute() (in CMDLINE mode)
command_line_changed() (in CMDLINE mode)
getexline() (used when executing from register)
openscript() (return value is not used)
vpeekc_any() which is used in
    ins_compl_check_keys() (in INSERT mode)
    f_getchar() (checked against empty (i.e. LANGMAP doesn't matter)
    command_line_changed() (check for empty)
char_avail() (checked against empty)
normal_get_additional_char() (Check for empty,
                              Check for empty,
                              Check for multibyte to determine
                              whether need to read more, any more is
                              read with plain_vgetc())
do_mouse() (Check for empty)

Before this commit, there are 7 exceptions to the rule that 'langmap' is
applied whenever neovim takes a key to be interpreted as a command.
These exceptions are mentioned in the test spec.
This commit changes two of these so they are no longer exceptions,
these are: confirming whether to apply a substitution after :s///c, and
asking y/n after a backwards range.

I'm leaving the changes as I believe they are improvements.
On the other hand, I'm not changing the behaviour of the other
exceptions despite being of the opinion that they should change, as the
point of this commit is to record the 'langmap' translation in macros,
not to change when the 'langmap' translation is applied.