Use structs and enums to specify data
Rather than pass values by reference into functions that only need to
update those values, we don't give functions that information, and
make all updates in the calling function by copying members from a
returned structure.
This makes it much clearer what information a function may be affected
by, and what information is updated.

Also use an enum in a few places, to limit the information a function is
given.