Fix folding problem with :move
In PR #6221 there was a mistake in calculating which folds need to be
re-ordered. When there are no folds after those that have been adjusted,
then the 'move_end' variable is left as 0.
This results in reverse_fold_order() swapping folds that have been
adjusted with uninitialised folds in the remainder of the grow array.

We add a check in foldMoveRange() to account for this case.