Fix up "linux*" terminfo entries, too.
The terminfo entry for linux only advertises 8 colours, but nvim
tries to make it display 16 colours anyway, resulting in erroneous
SGR control sequences for colours 8 and above.
The Linux kernel terminal emulator itself has actually understood
the 256-colour control sequences since version 4.8 and the
16-colour control sequences since version 4.9.
Thus we apply the same terminfo fixup as we apply for *xterm* and
*256*, to emit the 16-colour and 256-colour control sequences even
if terminfo's setaf and setab do not advertise them.