Revert 'pastetoggle': support value >1 char
Commit 337b6179df852350b52409fd3806e4b47ab2875b
This fixes issue #6716 at the expense of not being able to use a
multi-key 'pastetoggle' manually.
A multi-key 'pastetoggle' can still be used when inserting the entire
option into the typebuffer at once (though the use here is
questionable).

Also remove those tests to do with waiting for the completion of
'pastetoggle' and mention in the documentation that 'pastetoggle'
doesn't wait for timeout.