Reset the scroll region when resizing.
DECSLPP is explicitly documented as not affecting the scroll region.
The dtterm extension is not as well documented, but it is safer than not to assume that it operates similarly.

This also eliminates a pointlessly repeated test from tui_scroll().
It additionally uses a non-parameterized DECSTBM sequence when attempting to reset back to whole-screen scrolling.