Remove bashisms from os shell unit test.
bash is not universally available, and is not universally /bin/bash even where
it is available.  And it is not even required by this test in the first place.

/bin/sh serves the needs here just fine.  The Korn and Almquist shells, where
used for /bin/sh on operating systems such as OpenBSD, Debian, and FreeBSD,
both have the -n extension to echo that this test relies upon.