terminal: Do not change 'number', 'relativenumber'
Showing the 'number' column in terminal buffers is a bit silly because
of 'scrollback'. But it's mostly harmless and technically works as
expected. It also avoids "bounce" when going to/from terminal-mode.
So the least surprising thing is to respect the user's global setting.

We still disable 'relativenumber' in *terminal-mode* (as opposed to
normal-mode) because it is totally broken: the Nvim cursor (not terminal
cursor) is always on the last line.